+++ packages/menu-dropdown/src/MenuDropdownContent/index.style.ts
@@ -1,11 +1,12 @@
 import styled from 'styled-components';
 import { depth } from '@lendi-ui/depth';
 
-export interface ContentWrapperProps {
-  width: number;
-}
-
 export const ContentWrapper = styled.div`
+  position: absolute;
+  display: block;
+  width: 100%;
+  z-index: 1;
+  ${depth(3)};
   width: 250px;
   overflow: hidden;
   border-radius: 6px;