+++ packages/menu-dropdown/src/MenuDropdownContent/index.tsx
@@ -2,8 +2,10 @@ import * as React from 'react';
 import { MenuDropdownItem, ItemProps } from '../MenuDropdownItem';
 import { ContentWrapper } from './index.style';
 import { MenuDropdownDivider } from '../MenuDropdownDivider';
+import { LUIGlobalProps } from '@lendi-ui/utils';
+import MenuDropdownContext from '../MenuDropdownContext';
 
-export class MenuDropdownContent extends React.Component {
+export class MenuDropdownContent extends React.Component<LUIGlobalProps> {
   static Item = MenuDropdownItem;
   static Divider = MenuDropdownDivider;
 
@@ -15,7 +17,17 @@ export class MenuDropdownContent extends React.Component {
   };
 
   render() {
-    const { children } = this.props;
-    return <ContentWrapper>{React.Children.map(children as React.ReactChild, this.renderItem)}</ContentWrapper>;
+    const { children, ...otherProps } = this.props;
+    return (
+      <MenuDropdownContext.Consumer>
+        {({ showDropdown }) =>
+          showDropdown && (
+            <ContentWrapper {...otherProps}>
+              {React.Children.map(children as React.ReactChild, this.renderItem)}
+            </ContentWrapper>
+          )
+        }
+      </MenuDropdownContext.Consumer>
+    );
   }
 }