+++ packages/menu-dropdown/src/MenuDropdownDivider/index.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { Divider } from './index.style';
+import { LUIGlobalProps } from '@lendi-ui/utils';
 
-export const MenuDropdownDivider: React.FunctionComponent = () => {
-  return <Divider />;
+export const MenuDropdownDivider: React.FunctionComponent<LUIGlobalProps> = (props) => {
+  return <Divider {...props} />;
 };