+++ packages/menu-dropdown/src/index.tsx
@@ -1,1 +1,4 @@
-export { default } from './MenuDropdown';
+import MenuDropdown from './MenuDropdown';
+export * from './MenuDropdownContext';
+
+export default MenuDropdown;